Move `rpmmd.Load*Repositories()` to `reporegistry`
The `LoadAllRepositories()` will need to be extended with a
compatibility layer to eventually support loading distro configurations
with and without a dot in their name, under the same distro name. This
is to allow us to use dots in new distro config files, while not
breaking any existing repo overrides, that the user can have on their
system. IOW `rhel-84.json` and `rhel-8.4.json` should both correspond to
`rhel-8.4` distro. This will require using the common
`distroidstrparser` to parse distro ID string. However, the parser uses
distro-specific parsers, which results in an import loop when importing
`distroidstrparser` in `rpmmd` package.

To break this future import loop, let's move the `Load*Repositories()`
with their tests to the `reporegistry` package. This makes some sense,
since these are higher-level functions, than just those for loading the
actual data from the JSON file.

Make necessary adjustments in code which used these functions.

The `loadRepositoriesFromFile()` in `rpmmd` is now made public and it is
used to load a single repo config from a specific JSON file.
