+++ pkg/reporegistry/test/confpaths/priority1/repositories/test-distro.json
