+++ pkg/reporegistry/test/confpaths/priority2/repositories/test-distro.json
